Separate new and legacy versions of the crypt handle
