+++ src/engine/strat_engine/backstore/blockdevmgr.rs
@@ -22,7 +22,7 @@ use crate::{
                 devices::{initialize_devices, wipe_blockdevs, UnownedDevices},
                 shared::{BlkDevSegment, Segment},
             },
-            crypt::CryptHandle,
+            crypt::handle::v1::CryptHandle,
             metadata::{MDADataSize, BDA},
             serde_structs::{BaseBlockDevSave, Recordable},
             shared::bds_to_bdas,
