+++ src/engine/strat_engine/crypt/shared.rs
@@ -3,7 +3,6 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 use std::{
-    fmt::{self, Formatter},
     fs::OpenOptions,
     io::Write,
     mem::forget,
@@ -13,16 +12,11 @@ use std::{
 
 use data_encoding::BASE64URL_NOPAD;
 use either::Either;
-use serde::{
-    de::{Error, MapAccess, Visitor},
-    ser::SerializeMap,
-    Deserialize, Deserializer, Serialize, Serializer,
-};
-use serde_json::{from_value, to_value, Map, Value};
+use serde_json::{Map, Value};
 use sha2::{Digest, Sha256};
 use tempfile::TempDir;
 
-use devicemapper::{Bytes, DevId, DmName, DmNameBuf, DmOptions};
+use devicemapper::{Bytes, DevId, DmName, DmOptions};
 use libcryptsetup_rs::{
     c_uint,
     consts::{
@@ -31,33 +25,23 @@ use libcryptsetup_rs::{
             CryptDebugLevel, CryptLogLevel, CryptStatusInfo, CryptWipePattern, EncryptionFormat,
         },
     },
-    register, set_debug_level, set_log_callback, CryptDevice, CryptInit, TokenInput,
+    register, set_debug_level, set_log_callback, CryptDevice, CryptInit,
 };
 
 use crate::{
     engine::{
         strat_engine::{
-            backstore::get_devno_from_path,
             cmd::clevis_decrypt,
-            crypt::{
-                consts::{
-                    CLEVIS_LUKS_TOKEN_ID, CLEVIS_RECURSION_LIMIT, CLEVIS_TANG_TRUST_URL,
-                    CLEVIS_TOKEN_NAME, DEFAULT_CRYPT_KEYSLOTS_SIZE, DEFAULT_CRYPT_METADATA_SIZE,
-                    LUKS2_SECTOR_SIZE, LUKS2_TOKEN_ID, LUKS2_TOKEN_TYPE, STRATIS_TOKEN_DEVNAME_KEY,
-                    STRATIS_TOKEN_DEV_UUID_KEY, STRATIS_TOKEN_ID, STRATIS_TOKEN_POOLNAME_KEY,
-                    STRATIS_TOKEN_POOL_UUID_KEY, STRATIS_TOKEN_TYPE, TOKEN_KEYSLOTS_KEY,
-                    TOKEN_TYPE_KEY,
-                },
-                handle::{CryptHandle, CryptMetadata},
+            crypt::consts::{
+                CLEVIS_LUKS_TOKEN_ID, CLEVIS_RECURSION_LIMIT, CLEVIS_TANG_TRUST_URL,
+                CLEVIS_TOKEN_NAME, DEFAULT_CRYPT_KEYSLOTS_SIZE, DEFAULT_CRYPT_METADATA_SIZE,
+                LUKS2_SECTOR_SIZE, LUKS2_TOKEN_ID, LUKS2_TOKEN_TYPE, TOKEN_KEYSLOTS_KEY,
+                TOKEN_TYPE_KEY,
             },
-            dm::{get_dm, DEVICEMAPPER_PATH},
+            dm::get_dm,
             keys,
-            metadata::StratisIdentifiers,
         },
-        types::{
-            DevUuid, DevicePath, EncryptionInfo, KeyDescription, Name, PoolUuid, SizedKeyMemory,
-            UnlockMethod,
-        },
+        types::{KeyDescription, SizedKeyMemory, UnlockMethod},
     },
     stratis::{StratisError, StratisResult},
 };
@@ -78,185 +62,6 @@ pub fn set_up_crypt_logging() {
 
     set_debug_level(CryptDebugLevel::All);
     set_log_callback::<()>(Some(c_logging_callback), None);
-}
-
-pub struct StratisLuks2Token {
-    pub devname: DmNameBuf,
-    pub identifiers: StratisIdentifiers,
-    pub pool_name: Option<Name>,
-}
-
-impl Serialize for StratisLuks2Token {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let mut map_serializer = serializer.serialize_map(None)?;
-        map_serializer.serialize_entry(TOKEN_TYPE_KEY, STRATIS_TOKEN_TYPE)?;
-        map_serializer.serialize_entry::<_, [u32; 0]>(TOKEN_KEYSLOTS_KEY, &[])?;
-        map_serializer.serialize_entry(STRATIS_TOKEN_DEVNAME_KEY, &self.devname.to_string())?;
-        map_serializer.serialize_entry(
-            STRATIS_TOKEN_POOL_UUID_KEY,
-            &self.identifiers.pool_uuid.to_string(),
-        )?;
-        map_serializer.serialize_entry(
-            STRATIS_TOKEN_DEV_UUID_KEY,
-            &self.identifiers.device_uuid.to_string(),
-        )?;
-        if let Some(ref pn) = self.pool_name {
-            map_serializer.serialize_entry(STRATIS_TOKEN_POOLNAME_KEY, pn)?;
-        }
-        map_serializer.end()
-    }
-}
-
-impl<'de> Deserialize<'de> for StratisLuks2Token {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct StratisTokenVisitor;
-
-        impl<'de> Visitor<'de> for StratisTokenVisitor {
-            type Value = StratisLuks2Token;
-
-            fn expecting(&self, f: &mut Formatter<'_>) -> fmt::Result {
-                write!(f, "a Stratis LUKS2 token")
-            }
-
-            fn visit_map<A>(self, mut map: A) -> Result<Self::Value, A::Error>
-            where
-                A: MapAccess<'de>,
-            {
-                let mut token_type = None;
-                let mut token_keyslots = None;
-                let mut d_name = None;
-                let mut p_uuid = None;
-                let mut d_uuid = None;
-                let mut p_name = None;
-
-                while let Some((k, v)) = map.next_entry::<String, Value>()? {
-                    match k.as_str() {
-                        TOKEN_TYPE_KEY => {
-                            token_type = Some(v);
-                        }
-                        TOKEN_KEYSLOTS_KEY => {
-                            token_keyslots = Some(v);
-                        }
-                        STRATIS_TOKEN_DEVNAME_KEY => {
-                            d_name = Some(v);
-                        }
-                        STRATIS_TOKEN_POOL_UUID_KEY => {
-                            p_uuid = Some(v);
-                        }
-                        STRATIS_TOKEN_DEV_UUID_KEY => {
-                            d_uuid = Some(v);
-                        }
-                        STRATIS_TOKEN_POOLNAME_KEY => {
-                            p_name = Some(v);
-                        }
-                        st => {
-                            return Err(A::Error::custom(format!("Found unrecognized key {st}")));
-                        }
-                    }
-                }
-
-                token_type
-                    .ok_or_else(|| A::Error::custom(format!("Missing field {TOKEN_TYPE_KEY}")))
-                    .and_then(|ty| match ty {
-                        Value::String(s) => {
-                            if s == STRATIS_TOKEN_TYPE {
-                                Ok(())
-                            } else {
-                                Err(A::Error::custom(format!(
-                                    "Incorrect value for {TOKEN_TYPE_KEY}: {s}"
-                                )))
-                            }
-                        }
-                        _ => Err(A::Error::custom(format!(
-                            "Unrecognized value type for {TOKEN_TYPE_KEY}"
-                        ))),
-                    })
-                    .and_then(|_| {
-                        let value = token_keyslots.ok_or_else(|| {
-                            A::Error::custom(format!("Missing field {TOKEN_KEYSLOTS_KEY}"))
-                        })?;
-                        match value {
-                            Value::Array(a) => {
-                                if a.is_empty() {
-                                    Ok(())
-                                } else {
-                                    Err(A::Error::custom(format!(
-                                        "Found non-empty array for {TOKEN_KEYSLOTS_KEY}"
-                                    )))
-                                }
-                            }
-                            _ => Err(A::Error::custom(format!(
-                                "Unrecognized value type for {TOKEN_TYPE_KEY}"
-                            ))),
-                        }
-                    })
-                    .and_then(|_| {
-                        let value = d_name.ok_or_else(|| {
-                            A::Error::custom(format!("Missing field {STRATIS_TOKEN_DEVNAME_KEY}"))
-                        })?;
-                        match value {
-                            Value::String(s) => DmNameBuf::new(s).map_err(A::Error::custom),
-                            _ => Err(A::Error::custom(format!(
-                                "Unrecognized value type for {STRATIS_TOKEN_DEVNAME_KEY}"
-                            ))),
-                        }
-                    })
-                    .and_then(|dev_name| {
-                        let value = p_uuid.ok_or_else(|| {
-                            A::Error::custom(format!("Missing field {STRATIS_TOKEN_POOL_UUID_KEY}"))
-                        })?;
-                        match value {
-                            Value::String(s) => PoolUuid::parse_str(&s)
-                                .map(|uuid| (dev_name, uuid))
-                                .map_err(A::Error::custom),
-                            _ => Err(A::Error::custom(format!(
-                                "Unrecognized value type for {STRATIS_TOKEN_POOL_UUID_KEY}"
-                            ))),
-                        }
-                    })
-                    .and_then(|(dev_name, pool_uuid)| {
-                        let value = d_uuid.ok_or_else(|| {
-                            A::Error::custom(format!("Missing field {STRATIS_TOKEN_DEV_UUID_KEY}"))
-                        })?;
-                        match value {
-                            Value::String(s) => DevUuid::parse_str(&s)
-                                .map(|uuid| (dev_name, pool_uuid, uuid))
-                                .map_err(A::Error::custom),
-                            _ => Err(A::Error::custom(format!(
-                                "Unrecognized value type for {STRATIS_TOKEN_DEV_UUID_KEY}"
-                            ))),
-                        }
-                    })
-                    .and_then(|(devname, pool_uuid, device_uuid)| {
-                        let pool_name = match p_name {
-                            Some(Value::String(s)) => Some(Name::new(s)),
-                            Some(_) => {
-                                return Err(A::Error::custom(format!(
-                                    "Unrecognized value type for {STRATIS_TOKEN_POOLNAME_KEY}"
-                                )))
-                            }
-                            None => None,
-                        };
-                        Ok(StratisLuks2Token {
-                            devname,
-                            identifiers: StratisIdentifiers {
-                                pool_uuid,
-                                device_uuid,
-                            },
-                            pool_name,
-                        })
-                    })
-            }
-        }
-
-        deserializer.deserialize_map(StratisTokenVisitor)
-    }
 }
 
 /// Acquire a crypt device handle or return an error. This serves as a wrapper
@@ -350,121 +155,9 @@ pub fn add_keyring_keyslot(
     Ok(())
 }
 
-/// Set up a libcryptsetup device handle on a device that may or may not be a LUKS2
-/// device.
-pub fn setup_crypt_device(physical_path: &Path) -> StratisResult<Option<CryptDevice>> {
-    let device_result = device_from_physical_path(physical_path);
-    match device_result {
-        Ok(None) => Ok(None),
-        Ok(Some(mut dev)) => {
-            if !is_encrypted_stratis_device(&mut dev) {
-                Ok(None)
-            } else {
-                Ok(Some(dev))
-            }
-        }
-        Err(e) => Err(e),
-    }
-}
-
-/// Load crypt device metadata.
-pub fn load_crypt_metadata(
-    device: &mut CryptDevice,
-    physical_path: &Path,
-) -> StratisResult<Option<CryptMetadata>> {
-    let physical = DevicePath::new(physical_path)?;
-
-    let identifiers = identifiers_from_metadata(device)?;
-    let activation_name = activation_name_from_metadata(device)?;
-    let pool_name = pool_name_from_metadata(device)?;
-    let key_description = key_desc_from_metadata(device);
-    let devno = get_devno_from_path(physical_path)?;
-    let key_description = match key_description
-        .as_ref()
-        .map(|kd| KeyDescription::from_system_key_desc(kd))
-    {
-        Some(Some(Ok(description))) => Some(description),
-        Some(Some(Err(e))) => {
-            return Err(StratisError::Msg(format!(
-                "key description {} found on devnode {} is not a valid Stratis key description: {}",
-                key_description.expect("key_desc_from_metadata determined to be Some(_) above"),
-                physical_path.display(),
-                e,
-            )));
-        }
-        Some(None) => {
-            warn!("Key description stored on device {} does not appear to be a Stratis key description; ignoring", physical_path.display());
-            None
-        }
-        None => None,
-    };
-    let clevis_info = clevis_info_from_metadata(device)?;
-
-    let encryption_info =
-        if let Some(info) = EncryptionInfo::from_options((key_description, clevis_info)) {
-            info
-        } else {
-            return Err(StratisError::Msg(format!(
-                "No valid encryption method that can be used to unlock device {} found",
-                physical_path.display()
-            )));
-        };
-
-    let path = vec![DEVICEMAPPER_PATH, &activation_name.to_string()]
-        .into_iter()
-        .collect::<PathBuf>();
-    let activated_path = path.canonicalize().unwrap_or(path);
-    Ok(Some(CryptMetadata {
-        physical_path: physical,
-        identifiers,
-        encryption_info,
-        activation_name,
-        pool_name,
-        device: devno,
-        activated_path,
-    }))
-}
-
-/// Set up a handle to a crypt device using either Clevis or the keyring to activate
-/// the device.
-pub fn setup_crypt_handle(
-    device: &mut CryptDevice,
-    physical_path: &Path,
-    unlock_method: Option<UnlockMethod>,
-) -> StratisResult<Option<CryptHandle>> {
-    let metadata = match load_crypt_metadata(device, physical_path)? {
-        Some(m) => m,
-        None => return Ok(None),
-    };
-
-    if !vec![DEVICEMAPPER_PATH, &metadata.activation_name.to_string()]
-        .into_iter()
-        .collect::<PathBuf>()
-        .exists()
-    {
-        if let Some(unlock) = unlock_method {
-            activate(
-                device,
-                metadata.encryption_info.key_description(),
-                unlock,
-                &metadata.activation_name,
-            )?;
-        }
-    }
-
-    Ok(Some(CryptHandle::new(
-        metadata.physical_path,
-        metadata.identifiers.pool_uuid,
-        metadata.identifiers.device_uuid,
-        metadata.encryption_info,
-        metadata.pool_name,
-        metadata.device,
-    )))
-}
-
 /// Create a device handle and load the LUKS2 header into memory from
 /// a physical path.
-fn device_from_physical_path(physical_path: &Path) -> StratisResult<Option<CryptDevice>> {
+pub fn device_from_physical_path(physical_path: &Path) -> StratisResult<Option<CryptDevice>> {
     let mut device = log_on_failure!(
         CryptInit::init(physical_path),
         "Failed to acquire a context for device {}",
@@ -725,51 +418,6 @@ fn pin_dispatch(decoded_jwe: &Value, recursion_limit: u64) -> StratisResult<(Str
     }
 }
 
-/// Check whether the physical device path corresponds to an encrypted
-/// Stratis device.
-///
-/// This method works on activated and deactivated encrypted devices.
-///
-/// This device will only return true if the device was initialized
-/// with encryption by Stratis. This requires that:
-/// * the device is a LUKS2 encrypted device.
-/// * the device has a valid Stratis LUKS2 token.
-fn is_encrypted_stratis_device(device: &mut CryptDevice) -> bool {
-    fn device_operations(device: &mut CryptDevice) -> StratisResult<()> {
-        let stratis_token = device.token_handle().json_get(STRATIS_TOKEN_ID).ok();
-        let luks_token = device.token_handle().json_get(LUKS2_TOKEN_ID).ok();
-        let clevis_token = device.token_handle().json_get(CLEVIS_LUKS_TOKEN_ID).ok();
-        if stratis_token.is_none() || (luks_token.is_none() && clevis_token.is_none()) {
-            return Err(StratisError::Msg(
-                "Device appears to be missing some of the required Stratis LUKS2 tokens"
-                    .to_string(),
-            ));
-        }
-        if let Some(ref lt) = luks_token {
-            if !luks2_token_type_is_valid(lt) {
-                return Err(StratisError::Msg("LUKS2 token is invalid".to_string()));
-            }
-        }
-        if let Some(st) = stratis_token {
-            if !stratis_token_is_valid(st) {
-                return Err(StratisError::Msg("Stratis token is invalid".to_string()));
-            }
-        }
-        Ok(())
-    }
-
-    device_operations(device)
-        .map(|_| true)
-        .map_err(|e| {
-            debug!(
-                "Operations querying device to determine if it is a Stratis device \
-                failed with an error: {}; reporting as not a Stratis device.",
-                e
-            );
-        })
-        .unwrap_or(false)
-}
-
 fn device_is_active(device: Option<&mut CryptDevice>, device_name: &DmName) -> StratisResult<()> {
     match libcryptsetup_rs::status(device, &device_name.to_string()) {
         Ok(CryptStatusInfo::Active) => Ok(()),
@@ -815,7 +463,7 @@ pub fn activate(
     unlock_method: UnlockMethod,
     name: &DmName,
 ) -> StratisResult<()> {
-    if let (Some(kd), UnlockMethod::Keyring) = (key_desc, unlock_method) {
+    if let (Some(kd), UnlockMethod::Keyring | UnlockMethod::Any) = (key_desc, unlock_method) {
         let key_description_missing = keys::search_key_persistent(kd)
             .map_err(|_| {
                 StratisError::Msg(format!(
@@ -829,6 +477,8 @@ pub fn activate(
                 "Key description {} was not found in the keyring",
                 kd.as_application_str()
             );
+        }
+        if key_description_missing && unlock_method == UnlockMethod::Keyring {
             return Err(StratisError::Msg(format!(
                 "The key description \"{}\" is not currently set.",
                 kd.as_application_str(),
@@ -838,11 +488,13 @@ pub fn activate(
     log_on_failure!(
         device.token_handle().activate_by_token::<()>(
             Some(&name.to_string()),
-            Some(if unlock_method == UnlockMethod::Keyring {
-                LUKS2_TOKEN_ID
+            if unlock_method == UnlockMethod::Keyring {
+                Some(LUKS2_TOKEN_ID)
+            } else if unlock_method == UnlockMethod::Clevis {
+                Some(CLEVIS_LUKS_TOKEN_ID)
             } else {
-                CLEVIS_LUKS_TOKEN_ID
-            }),
+                None
+            },
             None,
             CryptActivate::empty(),
         ),
@@ -1048,26 +700,11 @@ pub fn check_luks2_token(device: &mut CryptDevice) -> StratisResult<()> {
 /// Validate that the LUKS2 token is present and valid
 ///
 /// May not be necessary. See the comment above the invocation.
-fn luks2_token_type_is_valid(json: &Value) -> bool {
+pub fn luks2_token_type_is_valid(json: &Value) -> bool {
     json.get(TOKEN_TYPE_KEY)
         .and_then(|type_val| type_val.as_str())
         .map(|type_str| type_str == LUKS2_TOKEN_TYPE)
         .unwrap_or(false)
-}
-
-/// Validate that the Stratis token is present and valid
-fn stratis_token_is_valid(json: Value) -> bool {
-    debug!("Stratis LUKS2 token: {}", json);
-
-    let result = from_value::<StratisLuks2Token>(json);
-    if let Err(ref e) = result {
-        debug!(
-            "LUKS2 token in the Stratis token slot does not appear \
-            to be a Stratis token: {}.",
-            e,
-        );
-    }
-    result.is_ok()
 }
 
 /// Read key from keyring with the given key description.
@@ -1090,43 +727,10 @@ pub fn read_key(key_description: &KeyDescription) -> StratisResult<Option<SizedK
     read_key_result.map(|opt| opt.map(|(_, mem)| mem))
 }
 
-/// Query the Stratis metadata for the device activation name.
-fn activation_name_from_metadata(device: &mut CryptDevice) -> StratisResult<DmNameBuf> {
-    Ok(from_value::<StratisLuks2Token>(device.token_handle().json_get(STRATIS_TOKEN_ID)?)?.devname)
-}
-
 /// Query the Stratis metadata for the key description used to unlock the
 /// physical device.
 pub fn key_desc_from_metadata(device: &mut CryptDevice) -> Option<String> {
     device.token_handle().luks2_keyring_get(LUKS2_TOKEN_ID).ok()
-}
-
-/// Query the Stratis metadata for the pool name.
-pub fn pool_name_from_metadata(device: &mut CryptDevice) -> StratisResult<Option<Name>> {
-    Ok(
-        from_value::<StratisLuks2Token>(device.token_handle().json_get(STRATIS_TOKEN_ID)?)?
-            .pool_name,
-    )
-}
-
-/// Replace the old pool name in the Stratis LUKS2 token.
-pub fn replace_pool_name(device: &mut CryptDevice, new_name: Name) -> StratisResult<()> {
-    let mut token =
-        from_value::<StratisLuks2Token>(device.token_handle().json_get(STRATIS_TOKEN_ID)?)?;
-    token.pool_name = Some(new_name);
-    device.token_handle().json_set(TokenInput::ReplaceToken(
-        STRATIS_TOKEN_ID,
-        &to_value(token)?,
-    ))?;
-    Ok(())
-}
-
-/// Query the Stratis metadata for the device identifiers.
-fn identifiers_from_metadata(device: &mut CryptDevice) -> StratisResult<StratisIdentifiers> {
-    Ok(
-        from_value::<StratisLuks2Token>(device.token_handle().json_get(STRATIS_TOKEN_ID)?)?
-            .identifiers,
-    )
 }
 
 // Bytes occupied by crypt metadata
