+++ src/engine/strat_engine/liminal/setup.rs
@@ -19,7 +19,7 @@ use crate::{
     engine::{
         strat_engine::{
             backstore::{StratBlockDev, UnderlyingDevice},
-            crypt::CryptHandle,
+            crypt::handle::v1::CryptHandle,
             device::blkdev_size,
             liminal::device_info::{LStratisDevInfo, LStratisInfo},
             metadata::BDA,
